fixing auth routes
