+++ src/app/login/auth-dialog/auth-dialog.component.html
@@ -9,22 +9,11 @@
       </div>
 
       <div class="row">
-        <h3 class="grey-text text-darken-2">{{authMode}}</h3>
-        <app-login-form (onFormResult)="onLoginFormResult($event)" *ngIf="isLoginMode()">
-        </app-login-form>
-        <app-register-form (onFormResult)="onRegisterFormResult($event)" *ngIf="isRegisterMode()">
+        <h3 class="grey-text text-darken-2">register</h3>
+        <app-register-form (onFormResult)="onRegisterFormResult($event)">
         </app-register-form>
       </div>
 
-      <div class="row">
-        <span class="grey-text text-darken-2" *ngIf="isLoginMode()">
-          Need an account? <a (click)="authMode='register'">Click here to register.</a>
-        </span>
-        <span class="grey-text text-darken-2" *ngIf="isRegisterMode()">
-          Already have an account? <a (click)="authMode='login'">Click here to login.</a>
-        </span>
-      </div>
-      
     </div>
   </div>
 </div>
