+++ src/app/login/auth-dialog/auth-dialog.component.ts
@@ -8,18 +8,9 @@ import { MaterializeAction } from "angular2-materialize";
 })
 export class AuthDialogComponent implements OnInit {
 
-  @Input('auth-mode') authMode: 'login' | 'register' = 'login';
   modalActions = new EventEmitter<string|MaterializeAction>();
 
   constructor() {}
-
-  onLoginFormResult(e){
-    if(e.signedIn)
-      this.closeDialog();
-    else{
-      alert(e.err.json().errors[0])
-    }
-  }
 
   onRegisterFormResult(e){
     if(e.signedUp)
@@ -29,8 +20,7 @@ export class AuthDialogComponent implements OnInit {
     }
   }
 
-  openDialog(mode: 'login' | 'register' = 'login'){
-    this.authMode = mode;
+  openDialog(){
     this.modalActions.emit({action:"modal", params:['open']});
   }
 
@@ -39,8 +29,5 @@ export class AuthDialogComponent implements OnInit {
   }
 
   ngOnInit() {}
-
-  isLoginMode(){return this.authMode == 'login'}
-  isRegisterMode(){return this.authMode == 'register'}
 
 }
