--- src/app/login/login-form/login-form.component.ts
@@ -1,41 +1,0 @@
-import { Component, OnInit, Output, EventEmitter } from '@angular/core';
-import { Router } from "@angular/router";
-import { AuthService } from "../../services/auth.service";
-
-@Component({
-  selector: 'app-login-form',
-  templateUrl: './login-form.component.html',
-  styleUrls: ['./login-form.component.css']
-})
-export class LoginFormComponent implements OnInit {
-
-  signInUser = {
-    email: '',
-    password: ''
-  };
-
-  @Output() onFormResult = new EventEmitter<any>();
-
-  constructor(public authService:AuthService,
-    private router: Router) { }
-
-  ngOnInit() {}
-
-  onSignInSubmit(){
-
-    this.authService.logInUser(this.signInUser).subscribe(
-        res => {
-          if(res.status == 200){
-            this.onFormResult.emit({signedIn: true, res});
-            this.router.navigate(['/students']);
-          }
-        },
-        err => {
-          console.log('err:', err);
-          this.onFormResult.emit({signedIn: false, err});
-        }
-    );
-
-  }
-
-}
