+++ src/app/login/login.component.html
@@ -1,8 +1,25 @@
 <div class="row">
-  <div class="col s12">
-    <h1>IncluCare</h1>
-    <p>
-      Seja bem vindo ao IncluCare.
-    </p>
+  <div class="container">
+  
+    <h5>LOGIN</h5>
+    <form (ngSubmit)="onSignInSubmit()" #f="ngForm" >
+      <div class="row">
+
+        <div class="input-field col s12">
+          <input id="email" type="email" required name='email' [(ngModel)]="signInUser.email" class="validate">
+          <label for="email">Email</label>
+        </div>
+
+        <div class="input-field col s12">
+          <input id="password" type="password" required name='password' [(ngModel)]="signInUser.password" class="validate">
+          <label for="password">Password</label>
+        </div>
+
+        <div class="col s12 ">
+          <button type="submit" [disabled]="!(f.valid)" [ngClass]="{'disabled': !(f.valid)}" class="waves-effect waves-light btn blue right"> Login </button>
+        </div>
+
+      </div>
+    </form>
   </div>
 </div>
