+++ src/app/login/login.component.spec.ts
@@ -1,5 +1,9 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
+import { HttpModule }                       from '@angular/http';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { RouterTestingModule }              from '@angular/router/testing';
+import { Angular2TokenService }             from 'angular2-token';
+import { AuthService }                      from "../services/auth.service";
 import { LoginComponent } from './login.component';
 
 describe('LoginComponent', () => {
@@ -7,8 +11,21 @@ describe('LoginComponent', () => {
   let fixture: ComponentFixture<LoginComponent>;
 
   beforeEach(async(() => {
+    let tokenMock = jasmine.createSpyObj('tokenMock', ['validateToken', 'subscribe']);
+    tokenMock.validateToken.and.returnValue(tokenMock);
+
     TestBed.configureTestingModule({
-      declarations: [ LoginComponent ]
+      declarations: [ LoginComponent ],
+      imports: [
+        HttpModule,
+        FormsModule,
+        ReactiveFormsModule,
+        RouterTestingModule
+      ],
+      providers: [
+        AuthService,
+        { provide: Angular2TokenService, useValue: tokenMock }
+      ]
     })
     .compileComponents();
   }));
