+++ src/app/login/login.component.ts
@@ -1,4 +1,6 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, Output, EventEmitter } from '@angular/core';
+import { Router } from "@angular/router";
+import { AuthService } from "../services/auth.service";
 
 @Component({
   selector: 'app-login',
@@ -7,9 +9,33 @@ import { Component, OnInit } from '@angular/core';
 })
 export class LoginComponent implements OnInit {
 
-  constructor() { }
+  signInUser = {
+    email: '',
+    password: ''
+  };
 
-  ngOnInit() {
+  @Output() onFormResult = new EventEmitter<any>();
+
+  constructor(public authService:AuthService,
+    private router: Router) { }
+
+  ngOnInit() {}
+
+  onSignInSubmit(){
+
+    this.authService.logInUser(this.signInUser).subscribe(
+        res => {
+          if(res.status == 200){
+            this.onFormResult.emit({signedIn: true, res});
+            this.router.navigate(['/students']);
+          }
+        },
+        err => {
+          console.log('err:', err);
+          this.onFormResult.emit({signedIn: false, err});
+        }
+    );
+
   }
 
 }
