+++ src/app/toolbar/toolbar.component.html
@@ -1,12 +1,10 @@
 <div class="navbar-fixed">
-  <nav  role="navigation">
-    <div class="nav-wrapper blue ">
+  <nav role="navigation">
+    <div class="nav-wrapper blue">
       <a [routerLink]="['']" id="logo-container" class="brand-logo">IncluCare</a>
       <ul class="right hide-on-med-and-down">
 
-        <li *ngIf="!(authService.userSignedIn$ | async)"><a (click)="presentAuthDialog()" >LOGIN</a></li>
-        <li *ngIf="!(authService.userSignedIn$ | async)"><a (click)="presentAuthDialog('register')">REGISTER</a></li>
-
+        <li *ngIf="(authService.userSignedIn$ | async)"><a (click)="presentAuthDialog()">REGISTER</a></li>
         <li *ngIf="(authService.userSignedIn$ | async)"><a [routerLink]="['/students']" > STUDENTS </a></li>
         <li *ngIf="(authService.userSignedIn$ | async)"><a [routerLink]="['/profile']">PROFILE</a></li>
         <li *ngIf="(authService.userSignedIn$ | async)"><a (click)="logOut()">LOGOUT</a></li>
